Fix merging add/remove on same method.
Changes to methods are detected as both addition/removal and merged into
a single "change" entry in a separate step.  The code that searched for
the equivalent "remove" step only ever looked at the first removed
method with that name, ch caused the merging to change if several
methods with the same name were changed at once (for example several
overloads were precated at once).

Bug: 27582395
Change-Id: I9718709de4d5e70c291c92ca2801e922f98a39a1